Refactor/#129 admin balancegame hook 분리
